refactoring to add more endpoints
