+++ backend/src/main/java/com/abdi/cardiscover/controller/agency/CreateAgency.java
@@ -1,0 +1,5 @@
+package com.abdi.cardiscover.controller.agency;
+
+public class CreateAgency {
+    
+}
