+++ backend/src/main/java/com/abdi/cardiscover/controller/speciality/CreateSpeciality.java
@@ -1,0 +1,5 @@
+package com.abdi.cardiscover.controller.speciality;
+
+public class CreateSpeciality {
+    
+}
