+++ backend/src/main/java/com/abdi/cardiscover/controller/supplier/CreateSupplier.java
@@ -1,0 +1,5 @@
+package com.abdi.cardiscover.controller.supplier;
+
+public class CreateSupplier {
+    
+}
