+++ backend/src/main/java/com/abdi/cardiscover/entity/ColorEntity.java
@@ -1,9 +1,6 @@
 package com.abdi.cardiscover.entity;
 
 import java.util.List;
-
-import com.abdi.cardiscover.entity.car.CarEntity;
-
 import java.util.ArrayList;
 
 import jakarta.persistence.CascadeType;
