+++ backend/src/main/java/com/abdi/cardiscover/entity/LocationEntity.java
@@ -1,9 +1,7 @@
-package com.abdi.cardiscover.entity.location;
+package com.abdi.cardiscover.entity;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import com.abdi.cardiscover.entity.car.CarEntity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
