+++ backend/src/main/java/com/abdi/cardiscover/entity/ReservationEntity.java
@@ -4,8 +4,6 @@ import java.util.ArrayList;
 import java.util.GregorianCalendar;
 import java.util.List;
 
-import com.abdi.cardiscover.entity.car.CarEntity;
-
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
