+++ backend/src/main/java/com/abdi/cardiscover/entity/SizeEntity.java
@@ -2,9 +2,6 @@ package com.abdi.cardiscover.entity;
 
 
 import java.util.List;
-
-import com.abdi.cardiscover.entity.brand.BrandEntity;
-
 import java.util.ArrayList;
 
 import jakarta.persistence.CascadeType;
