+++ backend/src/main/java/com/abdi/cardiscover/entity/SupplierEntity.java
@@ -3,8 +3,6 @@ package com.abdi.cardiscover.entity;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.abdi.cardiscover.entity.car.CarEntity;
-
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
