+++ backend/src/main/java/com/abdi/cardiscover/repository/BrandRepository.java
@@ -3,7 +3,7 @@ package com.abdi.cardiscover.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.abdi.cardiscover.entity.brand.BrandEntity;
+import com.abdi.cardiscover.entity.BrandEntity;
 
 @Repository
 public interface BrandRepository extends CrudRepository<BrandEntity, Long> {
