+++ backend/src/main/java/com/abdi/cardiscover/repository/LocationRepository.java
@@ -5,7 +5,7 @@ package com.abdi.cardiscover.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.abdi.cardiscover.entity.location.LocationEntity;
+import com.abdi.cardiscover.entity.LocationEntity;
 
 @Repository
 public interface LocationRepository extends CrudRepository<LocationEntity, Long> {
