+++ backend/src/main/java/com/abdi/cardiscover/requestbody/BrandRequestBody.java
@@ -1,0 +1,11 @@
+package com.abdi.cardiscover.requestbody;
+
+import java.math.BigDecimal;
+import java.util.GregorianCalendar;
+
+import org.springframework.format.annotation.DateTimeFormat;
+
+// This class maps JSON POST objects to Java objects via the @RequestBody  
+public class BrandRequestBody {
+  
+}
